* expr.c (store_constructor): SIZE now signed. 	For EXPR_SIZE, don't evaluate size; just needed if constant. 	* fold-const.c (fold): Fix a number of cases when folded tree is 	wrong type. 	* function.c (flush_addressof): Reenable. 	* tree.h (flush_addressof): Likewise. 	* toplev.c (independent_decode_option): Look at strings_processed. 	* config/alpha/alpha.h (MINIMUM_ATOMIC_ALIGNMENT): Cast to unsigned.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@32783 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/config/alpha/alpha.h
@@ -519,7 +519,7 @@ extern const char *alpha_mlat_string;	/* For -mmemory-latency= */
 
 /* For atomic access to objects, must have at least 32-bit alignment
    unless the machine has byte operations.  */
-#define MINIMUM_ATOMIC_ALIGNMENT (TARGET_BWX ? 8 : 32)
+#define MINIMUM_ATOMIC_ALIGNMENT ((unsigned int) (TARGET_BWX ? 8 : 32))
 
 /* Align all constants and variables to at least a word boundary so
    we can pick up pieces of them faster.  */
